Move DSN instance creation to the higher level
